pyproject.toml
[tool.aerich]
tortoise_orm = "db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"

[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["kiortir <mpgg2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = {extras = ["accel", "asyncpg"], version = "^0.19.3"}
aiogram = "^2.25.1"
pydantic = "^2.0"
redis = "^4.6.0"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
aerich = "^0.7.1"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true